don't let NG DNS-lookup hostnames itself...
... which it would do outside of the VPN anyway, and my router would
drop the packets.

Turning on or off the "Show log" button would restart the
VPN(apparently) and thus cause the port forwarding of port 53 to work
again because it'd stop after a while, this might be why it stopped.

Solved by assistance from Google's Gemini 2.5 Flash (has implicit
"Thinking" now)

This isn't tested yet, 25th of May 2025.
so I might be wrong about it fixing the need to do "Show log" as a
mitigation.
Yep, I'm wrong, it didn't fix the issue, mitigation via "Show log" is
still needed!
so, what did fix it and thus worked was removing the hostname
dns.nextdns.io from the personalDNSfilter app list within NetGuard, it
re-adds it as it's sensed via SNI but it never again went red after a
while; i conclude that killing NG in the past (once) might've screwed up
something in the db and removing that host reset or refreshed something
so it doesn't mess it up again by making it go red after a while.